+++ src/fastcgi/fastcgi_resource.cc
@@ -48,8 +48,8 @@ void onConnection(const TcpConnectionPtr& conn)
 
 void start_fastcgi()
 {
-  size_t port = get_config->m_port;
-  size_t threads = get_config->m_threads;
+  int port = get_config->m_port;
+  int threads = get_config->m_threads;
   LOG_INFO <<port<<":"<<threads;
   InetAddress addr(static_cast<uint16_t>(port));
   LOG_INFO << "FastCGI listens on " << addr.toIpPort()
